+++ Gruntfile.js
@@ -51,10 +51,6 @@ module.exports = function (grunt) {
       tasks: [
         'default'
       ]
-    },
-
-    emu: {
-      'README.md': 'index.js'
     }
 
   });
@@ -65,16 +61,7 @@ module.exports = function (grunt) {
   grunt.loadNpmTasks('grunt-contrib-uglify');
   grunt.loadNpmTasks('grunt-contrib-watch');
 
-  // tasks
-  grunt.registerMultiTask('emu', function () {
-    var emu = require('emu'),
-      fs = require('fs'),
-      source = fs.readFileSync(this.data, 'utf8');
-
-    fs.writeFileSync(this.target, emu.getComments(source));
-  });
-  grunt.registerTask('doc', ['emu']);
   grunt.registerTask('test', ['mochaTest']);
   grunt.registerTask('default', ['test', 'watch']);
-  grunt.registerTask('build', ['jshint', 'test', 'uglify', 'emu']);
+  grunt.registerTask('build', ['jshint', 'test', 'uglify']);
 };