+++ index.js
@@ -1,188 +1,3 @@
-/**
-    % tcomb
-
-    ![tcomb logo](http://gcanti.github.io/resources/tcomb/logo.png)
-
-    tcomb is a library for Node.js and the browser (2K gzipped) which allows you to **check the types** of 
-    JavaScript values at runtime with a simple syntax. It's great for **Domain Driven Design**, for checking external input, 
-    for testing and for adding safety to your internal code. 
-
-    # Feedback
-
-    This library aims to be a **pragmatic** tool for development, if you have any feedback that can improve
-    its usability, please let me know.
-
-    # Contents
-
-    - [Features](#features)
-    - [Quick Examples](#quick-examples)
-    - [Setup](#setup)
-    - [Requirements](#requirements)
-    - [Tests](#tests)
-    - [The Idea](#the-idea)
-    - [Api](#api)
-
-    ## Features
-
-    - **write complex domain models** in a breeze and with small code footprint
-    - easy debugging
-    - instances are immutables by default
-    - encode/decode of domain models to/from JSON for free
-
-    The library provides a built-in `assert` function, if an assert fails the **debugger kicks in** 
-    so you can inspect the stack and quickly find out what's wrong.
-
-    You can handle:
-
-    **JavaScript native types**
-
-    - Nil: `null` and `undefined`
-    - Str: strings
-    - Num: numbers
-    - Bool: booleans
-    - Arr: arrays
-    - Obj: plain objects
-    - Func: functions
-    - Err: errors
-    - Re: regular expressions
-    - Dat: dates
-    - Any: *
-
-    **type combinators** (build new types from those already defined)
-
-    - struct (i.e. classes)
-    - union
-    - maybe
-    - enums
-    - tuple
-    - subtype
-    - list
-    - dict
-    - function type (experimental)
-
-    ## Quick Examples
-
-    Let's build a product model
-
-    ```javascript
-    var Product = struct({
-        name: Str,                  // required string
-        desc: maybe(Str),           // optional string, can be null
-        home: Url,                  // a subtype of a string
-        shippings: list(Str),       // a list of shipping methods
-        category: Category,         // enum, one of [audio, video]
-        price: union(Num, Price),   // a price (dollars) OR in another currency
-        size: tuple([Num, Num]),    // width x height
-        warranty: dict(Num)         // a dictionary country -> covered years
-    });
-
-    var Url = subtype(Str, function (s) {
-        return s.indexOf('http://') === 0;
-    });
-
-    var Category = enums({ audio: 0, video: 1 });
-
-    var Price = struct({ currency: Str, amount: Num });
-
-    // JSON of a product
-    var json = {
-        name: 'iPod',
-        desc: 'Engineered for maximum funness.',
-        home: 'http://www.apple.com/ipod/',
-        shippings: ['Same Day', 'Next Businness Day'],
-        category: 'audio',
-        price: {currency: 'EUR', amount: 100},
-        size: [2.4, 4.1],
-        warranty: {
-          US: 2,
-          IT: 1
-        }
-    };
-
-    // get an immutable instance, `new` is optional
-    var ipod = Product(json);
-    ```
-
-    You have existing code and you want to add safety
-
-    ```javascript
-    // your code: plain old JavaScript class
-    function Point (x, y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    var p = new Point(1, 'a'); // silent error
-    ```
-
-    in order to "tcombify" your code you can simply add some asserts
-
-    ```javascript
-    function Point (x, y) {
-        assert(Num.is(x));
-        assert(Num.is(y));
-        this.x = x;
-        this.y = y;
-    }
-
-    var p = new Point(1, 'a'); // => fail! debugger kicks in
-    ```
-
-    ## Setup
-
-    Node
-
-        npm install tcomb
-
-    Browser
-
-        bower install tcomb
-
-    or download the `build/tcomb.min.js` file.
-
-    ## Requirements
-
-    This library uses a few ES5 methods
-
-    - `Array.forEach()`
-    - `Array.map()`
-    - `Array.some()`
-    - `Array.every()`
-    - `Object.keys()`
-    - `Object.freeze()`
-    - `JSON.stringify()`
-
-    you can use `es5-shim`, `es5-sham` and `json2` to support old browsers
-
-    ```html
-    <!--[if lt IE 9]>
-    <script src="json2.js"></script>
-    <script src="es5-shim.min.js"></script>
-    <script src="es5-sham.min.js"></script>
-    <![endif]-->
-    <script type="text/javascript" src="tcomb.js"></script>
-    <script type="text/javascript">
-        console.log(t);
-    </script>
-    ```
-
-    ## Tests
-
-    Run `mocha` or `npm test` in the project root.
-
-    ## The Idea
-
-    What's a type? In tcomb a type is a function `T` such that
-
-    1. `T` has signature `T(value, [mut])` where `value` depends on the nature of `T` and the optional boolean `mut` makes the instance mutable (default `false`)
-    2. `T` is idempotent: `T(T(value, mut), mut) === T(value, mut)`
-    3. `T` owns a static function `T.is(x)` returning `true` if `x` is an instance of `T`
-
-    **Note**: 2. implies that `T` can be used as a default JSON decoder
-
-    ## Api
-**/
-
 (function (root, factory) {
   'use strict';
   if (typeof define === 'function' && define.amd) {
@@ -196,46 +11,6 @@
 
   'use strict';
 
-  // rigger includes (https://github.com/buildjs/rigger)
-  // to view the full library code check out build/tcomb.js
-
-  /**
-      ### options
-  
-      #### function `options.onFail`
-  
-      In production envs you don't want to leak failures to the user
-  
-      ```javascript
-      // override onFail hook
-      options.onFail = function (message) {
-          try {
-              // capture stack trace
-              throw new Error(message);
-          } catch (e) {
-              // use you favourite JavaScript error logging service
-              console.log(e.stack);
-          }
-      };
-      ```
-  
-      #### function `options.update`
-  
-      Adds to structs, tuples, lists and dicts a static method `update` that returns a new instance
-      without modifying the original.
-  
-      Example
-  
-      ```javascript
-      // see http://facebook.github.io/react/docs/update.html
-      options.update = function (x, updates) {
-        return React.addons.update(mixin({}, x), updates);
-      };
-      var p1  = Point({x: 0, y: 0});
-      var p2 = Point.update(p1, {x: {$set: 1}}); // => Point({x: 1, y: 0})
-      ```
-  **/
-  
   var failed = false;
   
   function onFail(message) {
@@ -253,25 +28,6 @@
     update: null
   };
 
-  /**
-      ### assert(guard, [message], [values...]);
-  
-      If `guard !== true` the debugger kicks in.
-  
-      - `guard` boolean condition
-      - `message` optional string useful for debugging, formatted with values like [util.format in Node](http://nodejs.org/api/util.html#util_util_format_format)
-  
-      Example
-  
-      ```javascript
-      assert(1 === 2); // throws 'assert(): failed'
-      assert(1 === 2, 'error!'); // throws 'error!'
-      assert(1 === 2, 'error: %s !== %s', 1, 2); // throws 'error: 1 !== 2'
-      ```
-  
-      To customize failure behaviour, see `options.onFail`.
-  **/
-  
   function fail(message) {
     options.onFail(message);
   }
@@ -354,7 +110,26 @@
     assert(isType(type), 'Invalid argument `type` of value `%j` supplied to `getName()`, expected a type.', type);
     return type.meta.name;
   }
+
+  function getKind(type) {
+    assert(isType(type), 'Invalid argument `type` of value `%j` supplied to `geKind()`, expected a type.', type);
+    return type.meta.kind;
+  }
+
+  function isKind(type, kind) {
+    return getKind(type) === kind;
+  }
   
+  function values(obj) {
+    var ret = [];
+    for (var k in obj) {
+      if (obj.hasOwnProperty(k)) {
+        ret.push(obj[k]);
+      }
+    }
+    return ret;
+  }
+
   function ensureName(name, defaultName, types) {
     if (Nil.is(name)) {
       if (areTypes(types)) {
@@ -376,138 +151,83 @@
     assert(Func.is(options.update), errs.ERR_OPTIONS_UPDATE_MISSING);
     /*jshint validthis:true*/
     var T = this;
-    var args = slice.call(arguments);
-    var value = options.update.apply(options.update, args);
+    var value = options.update.apply(T, arguments);
     return T(value);
   }
 
   //
-  // Any - Because sometimes you really gonna need it.
+  // irriducibles
   //
   
-  function Any(value) {
-    forbidNewOperator(this, Any);
-    return value;
-  }
+  function irriducible(name, is) {
   
-  Any.meta = {
-    kind: 'any',
-    name: 'Any'
-  };
-  
-  Any.is = function () { return true; };
-
-  //
-  // primitives
-  //
-  
-  function primitive(name, is) {
-  
-    function Primitive(value) {
-      forbidNewOperator(this, Primitive);
+    function Irriducible(value) {
+      forbidNewOperator(this, Irriducible);
       assert(is(value), errs.ERR_BAD_TYPE_VALUE, value, name, format('a `%s`', name));
       // all primitives types are idempotent
       return value;
     }
   
-    Primitive.meta = {
-      kind: 'primitive',
+    Irriducible.meta = {
+      kind: 'irriducible',
       name: name
     };
   
-    Primitive.is = is;
+    Irriducible.is = is;
   
-    return Primitive;
+    return Irriducible;
   }
+
+  var Any = irriducible('Any', function () {
+    return true;
+  });
   
-  var Nil = primitive('Nil', function (x) {
+  var Nil = irriducible('Nil', function (x) {
     return x === null || x === undefined;
   });
   
-  var Str = primitive('Str', function (x) {
+  var Str = irriducible('Str', function (x) {
     return typeof x === 'string';
   });
   
-  var Num = primitive('Num', function (x) {
+  var Num = irriducible('Num', function (x) {
     return typeof x === 'number' && isFinite(x) && !isNaN(x);
   });
   
-  var Bool = primitive('Bool', function (x) {
+  var Bool = irriducible('Bool', function (x) {
     return x === true || x === false;
   });
   
-  var Arr = primitive('Arr', function (x) {
+  var Arr = irriducible('Arr', function (x) {
     return x instanceof Array;
   });
   
-  var Obj = primitive('Obj', function (x) {
+  var Obj = irriducible('Obj', function (x) {
     return !Nil.is(x) && typeof x === 'object' && !Arr.is(x);
   });
   
-  var Func = primitive('Func', function (x) {
+  var Func = irriducible('Func', function (x) {
     return typeof x === 'function';
   });
   
-  var Err = primitive('Err', function (x) {
+  var Err = irriducible('Err', function (x) {
     return x instanceof Error;
   });
   
-  var Re = primitive('Re', function (x) {
+  var Re = irriducible('Re', function (x) {
     return x instanceof RegExp;
   });
   
-  var Dat = primitive('Dat', function (x) {
+  var Dat = irriducible('Dat', function (x) {
     return x instanceof Date;
   });
-
-  /**
-      ### struct(props, [name])
-  
-      Defines a struct like type.
-  
-      - `props` hash name -> type
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      "use strict";
-  
-      // defines a struct with two numerical props
-      var Point = struct({
-          x: Num,
-          y: Num
-      });
-  
-      // methods are defined as usual
-      Point.prototype.toString = function () {
-          return '(' + this.x + ', ' + this.y + ')';
-      };
-  
-      // costructor usage, p is immutable
-      var p = Point({x: 1, y: 2});
-  
-      p.x = 2; // => TypeError
-  
-      p = Point({x: 1, y: 2}, true); // now p is mutable
-  
-      p.x = 2; // ok
-      ```
-  
-      #### is(x)
-  
-      Returns `true` if `x` is an instance of the struct.
-  
-      ```javascript
-      Point.is(p); // => true
-      ```
-  **/
   
   function struct(props, name) {
   
     // check combinator args
     name = ensureName(name, 'struct');
     assert(Obj.is(props), errs.ERR_BAD_COMBINATOR_ARGUMENT, 'props', props, name, 'an `Obj`');
+    assert(values(props).every(isType), errs.ERR_BAD_COMBINATOR_ARGUMENT, 'props', props, name, 'a dict of types');
   
     function Struct(value, mut) {
   
@@ -549,42 +269,6 @@
     return Struct;
   }
 
-  /**
-      ### union(types, [name])
-  
-      Defines a union of types.
-  
-      - `types` array of types
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      var Circle = struct({
-          center: Point,
-          radius: Num
-      });
-  
-      var Rectangle = struct({
-          bl: Point, // bottom left vertex
-          tr: Point  // top right vertex
-      });
-  
-      var Shape = union([
-          Circle, 
-          Rectangle
-      ]);
-      ```
-  
-      #### is(x)
-  
-      Returns `true` if `x` belongs to the union.
-  
-      ```javascript
-      Shape.is(Circle({center: p, radius: 10})); // => true
-      ```
-  **/
-  
   function union(types, name) {
   
     // check combinator args
@@ -596,7 +280,7 @@
       forbidNewOperator(this, Union);
       assert(Func.is(Union.dispatch), 'unimplemented %s.dispatch()', name);
       var T = Union.dispatch(value);
-      assert(isType(T), '%s.dispatch() returns no constructor', name);
+      assert(isType(T), '%s.dispatch() returns no type', name);
       // a union type is idempotent iif every T in types is idempotent
       return T(value, mut);
     }
@@ -613,24 +297,18 @@
       });
     };
   
+    // default dispatch implementation
+    Union.dispatch = function (x) {
+      for (var i = 0, len = types.length ; i < len ; i++ ) {
+        if (types[i].is(x)) {
+          return types[i];
+        }
+      }
+    };
+
     return Union;
   }
 
-  /**
-      ### maybe(type, [name])
-  
-      Same as `union([Nil, type])`.
-  
-      ```javascript
-      // the value of a radio input where null = no selection
-      var Radio = maybe(Str);
-  
-      Radio.is('a');     // => true
-      Radio.is(null);    // => true
-      Radio.is(1);       // => false
-      ```    
-  **/
-  
   function maybe(type, name) {
   
     // check combinator args
@@ -662,51 +340,6 @@
     return Maybe;
   }
 
-  /**
-      ### enums(map, [name])
-  
-      Defines an enum of strings.
-  
-      - `map` hash enum -> value
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      var Direction = enums({
-          North: 0, 
-          East: 1,
-          South: 2, 
-          West: 3
-      });
-      ```
-  
-      #### is(x)
-  
-      Returns `true` if `x` belongs to the enum.
-  
-      ```javascript
-      Direction.is('North'); // => true
-      ```
-      #### enums.of(keys, [name])
-  
-      Returns an enums of an array of keys, useful when you don't mind to define
-      custom values for the enums.
-  
-      - `keys` array (or string) of keys
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      // result is the same as the main example
-      var Direction = enums.of(['North', 'East', 'South', 'West']);
-  
-      // or..
-      Direction = enums.of('North East South West');
-      ```
-  **/
-  
   function enums(map, name) {
   
     // check combinator args
@@ -745,34 +378,6 @@
     return enums(value, name);
   };
 
-  /**
-      ### tuple(types, [name])
-  
-      Defines a tuple whose coordinates have the specified types.
-  
-      - `types` array of coordinates types
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      var Area = tuple([Num, Num]);
-  
-      // constructor usage, area is immutable
-      var area = Area([1, 2]);
-      ```
-  
-      #### is(x)
-  
-      Returns `true` if `x` belongs to the tuple.
-  
-      ```javascript
-      Area.is([1, 2]);      // => true
-      Area.is([1, 'a']);    // => false, the second element is not a Num
-      Area.is([1, 2, 3]);   // => false, too many elements
-      ```
-  **/
-  
   function tuple(types, name) {
   
     // check combinator args
@@ -829,44 +434,6 @@
     return Tuple;
   }
 
-  /**
-      ### subtype(type, predicate, [name])
-  
-      Defines a subtype of an existing type.
-  
-      - `type` the supertype
-      - `predicate` a function with signature `(x) -> boolean`
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      // points of the first quadrant
-      var Q1Point = subtype(Point, function (p) {
-          return p.x >= 0 && p.y >= 0;
-      });
-  
-      // costructor usage, p is immutable
-      var p = Q1Point({x: 1, y: 2});
-  
-      p = Q1Point({x: -1, y: -2}); // => fail!
-      ```
-      **Note**. You can't add methods to `Q1Point` `prototype`, add them to the supertype `prototype` if needed.
-  
-      #### is(x)
-  
-      Returns `true` if `x` belongs to the subtype.
-  
-      ```javascript
-      var Int = subtype(Num, function (n) {
-          return n === parseInt(n, 10);
-      });
-  
-      Int.is(2);      // => true
-      Int.is(1.1);    // => false
-      ```
-  **/
-  
   function subtype(type, predicate, name) {
   
     // check combinator args
@@ -907,37 +474,6 @@
     return Subtype;
   }
 
-  /**
-      ### list(type, [name])
-  
-      Defines an array where all the elements are of type `T`.
-  
-      - `type` type of all the elements
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      var Path = list(Point);
-  
-      // costructor usage, path is immutable
-      var path = Path([
-          {x: 0, y: 0}, 
-          {x: 1, y: 1}
-      ]);
-      ```
-  
-      #### is(x)
-  
-      Returns `true` if `x` belongs to the list.
-  
-      ```javascript
-      var p1 = Point({x: 0, y: 0});
-      var p2 = Point({x: 1, y: 2});
-      Path.is([p1, p2]); // => true
-      ```
-  **/
-  
   function list(type, name) {
   
     // check combinator args
@@ -990,30 +526,6 @@
     return List;
   }
 
-  /**
-      ### dict(type, [name])
-  
-      Defines a dictionary Str -> type.
-  
-      - `type` the type of the values
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      // defines a dictionary of numbers
-      var Tel = dict(Num);
-      ```
-  
-      #### is(x)
-  
-      Returns `true` if `x` is an instance of the dict.
-  
-      ```javascript
-      Tel.is({'jack': 4098, 'sape': 4139}); // => true
-      ```
-  **/
-
   function dict(type, name) {
   
     // check combinator args
@@ -1037,8 +549,7 @@
       var obj = {};
       for (var k in value) {
         if (value.hasOwnProperty(k)) {
-          var v = value[k];
-          obj[k] = type(v, mut);
+          obj[k] = type(value[k], mut);
         }
       }
   
@@ -1073,32 +584,11 @@
     return Dict;
   }
 
-  /**
-      ### func(Arguments, f, [Return], [name])
-  
-      **Experimental**. Defines a function where the `arguments` and the return value are checked.
-  
-      - `Arguments` the type of `arguments`
-      - `f` the function to execute
-      - `Return` optional, check the type of the return value
-      - `name` optional string useful for debugging
-  
-      Example
-  
-      ```javascript
-      var sum = func(tuple([Num, Num]), function (a, b) {
-          return a + b;
-      }, Num);
-  
-      sum(1, 2); // => 3
-      sum(1, 'a'); // => fail!
-      ```
-  **/
-  
   function func(Arguments, f, Return, name) {
   
     name = name || 'func()';
-    assert(isType(Arguments), errs.ERR_BAD_COMBINATOR_ARGUMENT, 'Arguments', Arguments, name, 'a type');
+    Arguments = Arr.is(Arguments) ? tuple(Arguments, 'Arguments') : Arguments;
+    assert(isType(Arguments), errs.ERR_BAD_COMBINATOR_ARGUMENT, 'Arguments', Arguments, name, 'a type or a list of types');
     assert(Func.is(f), errs.ERR_BAD_COMBINATOR_ARGUMENT, 'f', f, name, 'a `Func`');
     assert(Nil.is(Return) || isType(Return), errs.ERR_BAD_COMBINATOR_ARGUMENT, 'Return', Return, name, 'a type');
   
@@ -1144,14 +634,21 @@
   }
 
   return {
+
+    util: {
+      mixin: mixin,
+      format: format,
+      isType: isType,
+      getName: getName,
+      getKind: getKind,
+      isKind: isKind,
+      values: values,
+      slice: slice
+    },
+
     options: options,
     assert: assert,
-    mixin: mixin,
-    format: format,
-    isType: isType,
-    getName: getName,
     fail: fail,
-    slice: slice,
     
     Any: Any,
     Nil: Nil,
@@ -1165,6 +662,7 @@
     Re: Re,
     Dat: Dat,
 
+    irriducible: irriducible,
     struct: struct,
     enums: enums,
     union: union,
@@ -1176,39 +674,3 @@
     func: func
   };
 }));
-
-/**
-    ## IDEAS
-
-    - explore generating UI based on domain models written with tcomb
-    - explore auto validation of UI involving domain models written with tcomb
-    - explore using tcomb with React.js
-
-    ## Contribution
-
-    If you do have a contribution for the package feel free to put up a Pull Request or open an Issue.
-
-    ## License (MIT)
-
-    The MIT License (MIT)
-
-    Copyright (c) 2014 Giulio Canti
-
-    Permission is hereby granted, free of charge, to any person obtaining a copy
-    of this software and associated documentation files (the "Software"), to deal
-    in the Software without restriction, including without limitation the rights
-    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-    copies of the Software, and to permit persons to whom the Software is
-    furnished to do so, subject to the following conditions:
-
-    The above copyright notice and this permission notice shall be included in all
-    copies or substantial portions of the Software.
-
-    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-    SOFTWARE.
-**/